// src/common/enum/telecope.enum.ts
export enum TelescopeLocation {
    Space = "Space",
    Ground = "Ground",
}

export enum TelescopeSpectrum {
    Optical = "Optical",
    Infrared = "Infrared",
    Ultraviolet = "Ultraviolet",
    XRay = "XRay",
    GammaRay = "GammaRay",
    Radio = "Radio",
    Microwave = "Microwave",
    Submillimeter = "Submillimeter",
    Millimeter = "Millimeter",
    MultiWavelength = "MultiWavelength",
    Solar = "Solar",
    GravitationalWave = "GravitationalWave",
}